requirements.txt
alembic==1.11.1
anyio==3.6.2
astroid==2.15.4
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.95.1
fastapi-cli==0.0.5
feedparser==6.0.10
filelock==3.12.0
Flask==2.2.5
fonttools==4.38.0
geographiclib==2.0
geopy==2.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.11.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lxml==4.9.3
lz4==4.3.3
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mccabe==0.7.0
mdurl==0.1.2
node==1.2.1
numpy==1.21.6
odict==1.9.0
orjson==3.8.12
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
pipenv==2023.4.29
pipreqs==0.4.13
platformdirs==3.5.0
pluggy==1.2.0
plumber==1.7
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.18.0
pylint==2.17.4
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rich==13.7.1
rsa==4.9
rss-parser==0.2.4
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
sqlitecloud==0.0.78
starlette==0.26.1
swagger-ui-bundle==0.0.9
tomli==2.0.1
tomlkit==0.11.8
typer==0.12.3
typing_extensions==4.5.0
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
xmltodict==0.13.0
yarg==0.1.9
zipp==3.15.0
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
